+++ .env.sample
@@ -7,8 +7,11 @@ API_KEY=key-00000000000000000000000000000000
 # 3. Enter your publicly accessible hostname
 HOST=https://bouncebot.example.com
 
-# 4. (optional) Add a static CC to all bounce notifications
-#CC=helpdesk@example.com
+
+# Bounce config (optional)
+# Add a static CC to all bounce notifications
+#BOUNCE_CC=helpdesk@example.com
+
 
 # Mailing List config (optional)
 #MAILINGLIST_SUBJECT_PREFIX=[LISTSRV]
