+++ lib/strip-html.js
@@ -1,0 +1,9 @@
+const stripHtml = require('string-strip-html');
+
+module.exports.default = (html) => stripHtml(html, {
+    dumpLinkHrefsNearby: {
+        enabled: true,
+        wrapHeads: '(',
+        wrapTails: ')'
+    }
+});
