+++ package.json
@@ -14,8 +14,9 @@
   },
   "homepage": "https://github.com/anstosa/mailgun-bouncebot#readme",
   "dependencies": {
-    "mailgun-js": "^0.22.0",
     "body-parser": "^1.18.3",
-    "request": "^2.88.0"
+    "mailgun-js": "^0.22.0",
+    "request": "^2.88.0",
+    "string-strip-html": "^3.5.8"
   }
 }
